cambio tipo Long id

+++ src/main/java/uniandes/edu/co/proyecto/Modelo/Servicio.java
@@ -15,22 +15,22 @@ public class Servicio {
 
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)  
-    private Integer id;
+    private Long id;
     private String nombre;
 
     public Servicio()
     {;}
 
-    public Servicio(Integer id, String nombre) {
+    public Servicio(Long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
